ferrous_forge::cargo_intercept::wrapper

Function install_cargo_intercept

pub fn install_cargo_intercept() -> Result<()>

Install cargo interception system

Errors

Returns an error if the home directory cannot be determined, the install directory cannot be created, or the wrapper script cannot be written.